move some computation logic from view to model
